Add Tear Down stage and remove volumes.
